requirements.txt
pyyaml>=6.0
numpy>=1.24
scipy>=1.11
tqdm>=4.66
networkx>=3.2
metis==0.2a5
# --- PyTorch & PyG ---
torch>=2.3
torch_geometric>=2.5
torch_sparse>=0.6
torch_scatter>=2.1